fix `unnecessary cast` lint (#2596)
* fix clippy lints on Windows

* fix lints on other platforms

* a couple more

* again

* don't know what's goging on anymore

* fix examples

* comon

* how about now?

* this is getting annoying

* hmmm

* explicitly set a type

* :cry:

* don't cast on x64 targets

* apply code review requests

* fix attributes on expressions

* fix ios

+++ src/platform_impl/web/event_loop/runner.rs
@@ -322,8 +322,8 @@ impl<T: 'static> Shared<T> {
             };
             // First, we send the `ScaleFactorChanged` event:
             let current_size = crate::dpi::PhysicalSize {
-                width: canvas.width() as u32,
-                height: canvas.height() as u32,
+                width: canvas.width(),
+                height: canvas.height(),
             };
             let logical_size = current_size.to_logical::<f64>(old_scale);
             let mut new_size = logical_size.to_physical(new_scale);
