+++ src/platform_impl/windows/monitor.rs
@@ -241,7 +241,7 @@ impl MonitorHandle {
                 i += 1;
 
                 const REQUIRED_FIELDS: u32 =
-                    (DM_BITSPERPEL | DM_PELSWIDTH | DM_PELSHEIGHT | DM_DISPLAYFREQUENCY) as u32;
+                    DM_BITSPERPEL | DM_PELSWIDTH | DM_PELSHEIGHT | DM_DISPLAYFREQUENCY;
                 assert!(has_flag(mode.dmFields, REQUIRED_FIELDS));
 
                 // Use Ord impl of RootVideoMode
@@ -249,7 +249,7 @@ impl MonitorHandle {
                     video_mode: VideoMode {
                         size: (mode.dmPelsWidth, mode.dmPelsHeight),
                         bit_depth: mode.dmBitsPerPel as u16,
-                        refresh_rate_millihertz: mode.dmDisplayFrequency as u32 * 1000,
+                        refresh_rate_millihertz: mode.dmDisplayFrequency * 1000,
                         monitor: self.clone(),
                         native_video_mode: Box::new(mode),
                     },
